refactor(runtime/llm): extract compaction wrappers to own module
Moves the ~210 LOC compaction plumbing out of chat-executor-tool-loop.ts
into a new chat-executor-compaction-wrappers.ts. The main loop file
drops 2,219 -> 2,011 LOC.

Moved exports:
- runPerIterationCompactionBeforeModelCall (pre-call snip/microcompact/
  context-collapse/autocompact chain with PreCompact/PostCompact hook
  dispatch)
- callModelWithReactiveCompact (wraps a provider call so a 413 triggers
  an applyReactiveCompact drop + retry; seals open tool-protocol turns
  before the retry so the retry doesn't see dangling tool_call_ids)

extractCompactionLayerTag stays file-private in the new module.

Main loop imports dropped (now only used by the new module):
- applyPerIterationCompaction, computeAutocompactThreshold (./compact/index)
- applyReactiveCompact (./compact/reactive-compact)
- tryProjectedContextCollapse (./chat-executor-history-compaction)
- LLMContextWindowExceededError (./errors)

Purely file-organization — zero behavior delta. Typecheck green.
llm/ test suite (720 tests, 1 pre-existing token-budget compaction
failure unchanged from main) matches baseline.
